backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
black==26.3.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
distro==1.9.0
fastapi==0.136.1
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
isort==8.0.1
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.16
langchain-core==1.3.2
langchain-protocol==0.0.14
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.13
langsmith==0.7.38
mypy_extensions==1.1.0
openai==2.33.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.2
pathspec==1.1.1
platformdirs==4.9.6
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
python-multipart==0.0.27
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
xxhash==3.7.0
zstandard==0.25.0
